+++ CHANGELOG.md
@@ -34,8 +34,7 @@
     - OppositeVehicleRunningRedLight: Updated to allow execution at different locations
 * Added NHTSA Traffic Scenarios
     - ManeuverOppositeDirection: hero vehicle must maneuver in the opposite lane to pass a leading vehicle.
-* Added atomic behaviors using py_trees behavior trees library
-    - LeadingVehicleDecelerate: hero vehicle must react to the deceleration of leading vehicle and change lane to avoid collision and follow
+    - OtherLeadingVehicle: hero vehicle must react to the deceleration of leading vehicle and change lane to avoid collision and follow
       the vehicle in changed lane
 
 ## CARLA Scenario_Runner 0.9.2